ftp: Add perm to the default MLST/MLSD fact set
The facts (properties) listed by the MLST and MLSD commands in FTP are
configurable through the OPTS command. Most clients however do not both
checking the default facts of a server or to explicitly request a
different set of facts through the OPTS command.

Although this can easily be considered a bug in the client, it
nonetheless poses a compatibility problem. RFC 3659 lists perm amongst
the recommended minimum set of supported facts and it is conceivable
that clients interpreted this as perm being included in the default set.

Globus Online is one of the clients that expects perm to be returned
without explicitly requesting it.

We originally didn't include perm in the default set because it was
expensive to compute with the old permission handlers: the permission
handlers did additional round trips to pnfs manager for each file. This
is no longer a problem and this patch thus changes the default set to
include perm.

Target: trunk
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3456/