configuration: add obsolete removeUnexistingEntriesOnFlush
Way back in 1.7.0 era, the pool could take an property 
removeUnexistingEntriesOnFlush.  This property takes a boolean value 
and, if configured, would check if a file still exists immediately 
before flushing the file to tape.  If the file doesn't exist at that 
point then the file is deleted.

The current code-base does the same; however, this behaviour is 
always enabled.  This means that the property is no longer honoured.

This patch adds the property as an obsolete value with a short 
description of what's happening.

Target: trunk
Request: 1.9.12
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6401
Patch: http://rb.dcache.org/r/3459/
Acked-by: Tanja Baranova