dcap: fix overread when LCB is enabled
The LCB caching breaks a read into block-size chunks that it reads to 
satisfy some particular read request.  Once the byte-range of the 
request is satisfied, the read call returns.  Some client client 
(dccp, for example) don't bother to figure out how big is the 
requested file but simply keep reading a fixed sized block and 
terminate when a read returns zero.  This means that a read operation 
may request more data than is present in a file.

The current code-base doesn't cope with this possibility.  It notices 
that the final cache block has a short read, which it takes as some 
kind of problem.  As a fall-back mechanism, the code then disables LCB 
and satisfies the read request (and any subsequent ones) without 
caching.

The new code will attempt to read a second time after a short-read.  
If the second read is zero-length then EOF is assumed and the request 
is returned with all the data collected so-far.

Target: trunk
Patch: http://rb.dcache.org/r/3474/
Acked-by: Tigran Mkrtchyan