configuration: fix XML Entity handling of scoped assigments
dCache has the concept of scoped property value assignments.  A 
property assignment is considered scoped if the key has the form:

    <scope> "/" <name>

where <scope> is a non-empty string.

Scoped properties are used to define default values within the 
context of some specific service, identified by the string <scope>.  
Outside this context, the property assignment should have no effect.

The current implementation of ConfigurationProperties has no concept 
of scoped properties.  It's up to the code using 
ConfigurationProperties to decide if a property is scoped and, if so, 
what to do with it.

The XML Entity printer is outside any service scope, so all scoped 
property assignments should be ignored; however, the printer 
mistakenly includes all scoped properties in their "raw" form.

This patch tells the Entity printer to skip any property assignments 
that are scoped.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3468/
Acked-by: Gerd Behrmann