pool: throw a runtime exception if we fail to remove a file.
if some cases pool fail to remove data file but we accept
that situation. Even in case of a hardware problem if a pool
was unable to remove a file, the delete attempt was interpreted
as successful and cleaner entry was removed.

Ticket: 6248

Requested: 1.9.12, 1.9.5

Acked-by: Gerd Behrmann
Patch: http://rb.dcache.org/r/3482