migration: fix migrate-from-1.9.5.sh so it processes all of dCacheSetup
The migrate-from-1.9.5.sh script takes an existing dCache 
installation's 1.9.5 configuration and generates a "best-guess" 1.9.12 
configuration that gives the same behaviour.  If the site hasn't done 
anything too clever then this guess should be pretty good and, in any 
case, it should serve as a good starting point.  Part of the migration 
process adjusts the dCacheSetup file to generate a corresponding 
dcache.conf file.  This process is slightly non-trivial as duplicate 
assignments are not allowed.

The current version of the script will process only the first line of 
the dCacheSetup file for the dcache.conf file; all subsequent lines 
are discarded.

This patch fixes this problem.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3481/
Acked-by: Gerd Behrmann