ftp: Fix MLST/MLSD fact encoding
Since the UNIX.mode fact is not formally specified we are forced to
reverse engineer the encoding from existing implementations. Turns out
that both Globus Online and UberFTP expect the mode to be represented in
octal form.

The patch also adds authorization checks for the properties.

Finally, the patch suppresses the size fact for directory entries: The RFC
says that size is only for non-directory entries.

Target: trunk
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3460/