info-provider: fix info-provider for FHS-compliant build
The current info-provider for 1.9.12 (and trunk) is split over several 
files.  This avoids duplication but, more importantly, allows 
site-specific configuration to appear as a small file rather that 
being part of the larger file, as with previous releases.  There are 
several files that combine the different elements, including the 
site-specific configuration, into a single XML info-set.  These allow 
publishing of GLUE 1.3-only, GLUE 2.0-only and GLUE 1.3+2.0 
information.  However, these files need to locate the site-specific 
configuration in order to import the information.

The current info-provider assumes that the site-specific configuration 
is located at some constant relative path to the static information.  
This works for the /opt/d-cache deployments; however, with the 
FHS-compliant deployments, the offset isn't the same.  Also, it may be 
redefined later by a sysadmin, so the assumption isn't valid.

This patch fixes this problem by allowing the XML data to have access 
to dCache's configuration.  It does this by creating an additional 
file containing XML Entity definitions for all dCache properties with 
global scope.  The file is created as a temporary file (in /tmp) and, 
in order to avoid hard-coding the location, an XML catalogue is used 
to provide indirection to this file.  With this file, the location of 
the site-specific configuration file may be parametrised using the 
dcache.path.etc property; i.e., the &dcache.path.etc; XML Entity.

Target: trunk
Request: 1.9.12
Request: 1.9.13