xrootd: Fix transfer timeout
Xrootd transfers are handled by redirecting clients to pools. On the
pool movers implement the server part of the xrootd protocol. After
being redirected the door is no longer involved in the transfer.

The active transfers page in the monitoring system shows a list of
ongoing transfers in the system. That page is fed by information from
doors. For this reason xrootd doors have to maintain information about
xrootd movers even though they are not actually part of the transfer.

To deal with loss of the DoorTransferFinished message, xrootd doors have
a background task to clean stale information for movers that no longer
exist. The background task calls the finished() method on the Transfer
object when the mover is no longer found.

This fails to work in the xrootd door because the finished() method on
an XrootdTransfer doesn't remove the entry from the transfers map; that
was done in the message handler for the DoorTransferFinished message.
Thus stale entries could accumulate.

This patch moves the logic from the message handler and into the
finished method. This is the same approach used in the webdav door.

Target: trunk
Request: 1.9.13
Request: 1.9.12
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6421
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3489/