srm: move construction of contained requests inside ContainerRequest
The SRM specification includes several commands that operate on a list 
of files.  Support for holding information about these commands is 
provided by the various *Request classes (PutRequest, GetRequest, 
LsRequest, etc).  The individual files within a request are recorded as 
a FileRequest (or subclass thereof) objects.  Since all the *Request 
classes hold a list of files, the classes all extend a common 
superclass: ContainerRequest.

The ContainerRequest exposes the list of FileRequest objects to 
subclasses (and to the SrmPutDone handler) by making a field 
accessible.  It also requires that the subclass initialises this field 
"behind the back" of the superclass.  One of the side-effects of this 
is that it isn't terribly clear whether the FileRequest array has the 
possibility of being null.  This ambiguity is reflected in the code, 
where there is inconsistent support for the FileRequest array being 
null.

In fact, all subclasses initialise the FileRequest array in the 
abstract ContainerRequest so the FileRequest array is never null (at 
least, from what I can see!).  This means that the code testing for a 
null FileRequest array is unnecessary; however, the dead-code (or 
unused code-path of an "if" statement) fools FindBugs into thinking 
that SRM code may trigger a possible NPE.

One solution would be to simply remove all the tests for whether the 
FileRequest array is null.  This would be bad as doing so would leave 
the code brittle: a change in one of the constructors could result in 
the array potentially being null, triggering an NPE.

To solve this problem, this patch moves responsibility for initialising 
the list of FileRequest objects is moved into the ContainerRequest.  
The different subclass constructors are adjusted to produce a local 
array of the FileRequest objects that is passed to the ContainerRequest 
parent.

The patch also switches from using an array to using a array-backed 
List.  This should introduce no performance penalty but should allow 
future use of generics to cut down on duplicate code.

Target: trunk
Patch: http://rb.dcache.org/r/3478/
Acked-by: Gerd Behrmann
Acked-by: Dmitry Litvintsev