srm: refactor getNumOfFileRequest to collapse duplicated code
The SRM's ContainerRequest class is abstract.  One of the abstract 
methods is getNumOfFileRequests.  This is odd as the ContainerRequest 
class contains the FileRequest objects, so it should know how many 
there are!

In fact, the subclasses of ContainerRequest have different 
implementations.  Most simply count the number of requests but one of 
them (GetRequest) obtains a read-lock before counting the requests.  
Arguably, all should obtain a read-lock before counting.  This patch 
fixes this and also consolidates the implementations so that there is a 
single implementation of this method inside ContainerRequest.

Target: trunk
Patch: http://rb.dcache.org/r/3531/
Acked-by: Gerd Behrmann
Acked-by: Dmitry Litvintsev