CopyManager: register that the destination of the copy is writing data
Running the S2 tests, I observed the following stack-trace.

29 Jun 2011 18:44:56 (CopyManager) [v2:srmCopy:8662969:-2147472573:-2147472572 SRM-discordia CopyManager] Uncaught exception in thread Thread-155
java.lang.IllegalStateException: Can only delete name space entry for uploads
        at org.dcache.util.Transfer.deleteNameSpaceEntry(Transfer.java:864) ~[dcache.jar:na]
        at diskCacheV111.doors.CopyManager$CopyHandler.copy(CopyManager.java:446) ~[dcache.jar:na]
        at diskCacheV111.doors.CopyManager$CopyHandler.run(CopyManager.java:339) ~[dcache.jar:na]
        at java.lang.Thread.run(Thread.java:636) ~[na:1.6.0_20]

This appears to be triggered by an unsuccessful internal srmCopy 
command.  The error handling in CopyManager attempts to delete the 
namespace entry for the target file; however, since the target isn't 
registered as a write, the Transfer class refuses to delete the 
namespace entry.

This patch adjusts CopyManager so that it registers that the 
destination mover of the internal copy is writing a file.

Target: trunk
Request: 1.9.12
Request: 1.9.13
Patch: http://rb.dcache.org/r/3549/
Acked-by: Gerd Behrmann