dcap: don't report failed dcap operations as errors
The dcap door logs the communication with the dcap client as the 
result of the client issuing a command.  Operations that succeed 
are logged at debug level; those that fail are logged at error 
level.

It is wrong to log failed operations at error level as the door 
(and the rest of dCache) is functioning correctly.  For example, 
if the door has been instructed to open a non-existent directory 
for listing then the correct response is to return an error.  In 
doing so, the door (adn the rest of dCache) is behaving correctly.

One way of looking at the issue is that the door cannot know why 
the client made this request.  Perhaps the client expects the 
operation to fail, in which case a success should be logged at 
error-level.  Perhaps the client considers both success and 
failure as equally valid responses, so neither should be reported 
at error-level.

The case-in-point is that dccp will stat the destination before 
uploading the file.  When the destination doesn't yet exist 
(probably that vast majority of times) this results in an error 
being returned to the client.  However, it's worth reiterating 
that no error has happened here: the namespace, the door and the 
client have all functioned correctly, however an error is logged 
in dCache.  Since this happens for every uploaded file, the door's 
log file is awash with incorrect error messages, making real 
problems harder to see.

This patch adjusts the logging so that both successful and 
unsuccessful operations are logged at debug level.  Further 
information is made available to billing and, if a sysadmin needs 
to recover the messages sent to the client, they can enabled 
logging at debug level.

Target: trunk
Request: 1.9.13
Request: 1.9.12
Patch: http://rb.dcache.org/r/3557/
Acked-by: Tigran Mkrtchyan