srm: make Job.getJob type-safe
The SRM protocols features bulk-operations that allow multiple files to 
be processed in a similar fashion.  Within dCache's SRM implementation, 
the SRM API operations are represented by Requests objects and the 
individual files within a bulk-operation are FileRequests.  Both 
Requests and FileRequests are subclasses of the more generic Job class.

The Job class features static methods for fetching Jobs (either Request 
or FileRequest objects) from some storage: an in-memory cache, a shared 
memory cache or a database.  These methods return objects of some 
subclass of Job; however, often code wishes to manipulate the object 
based on its sub-class type.  This is handled in a somewhat 
inconsistent fashion: sometimes there is an explicit check and error 
handling, other times the code simply casts the object to the correct 
type, risking a ClassCastException.

This patch standardises and simplifies the process.  It introduces a 
new methods that wraps the existing method and casts the object to the 
correct form.  The existing getJob method is made private as there 
should be not direct (i.e., type unaware) access to this method.

Some points to note:

  o The Job ID may come from the SRM client, from the sys-admin or from 
    SRM internally.  Therefore, invoking getJob with an ID for the 
    wrong kind of Job isn't necessarily due to a programming error.

  o Both before and after the patch, getJob will throw 
    SRMInvalidRequestException if the job ID is unknown.

  o After the patch, getJob will throw SRMInvalidRequestException if 
    the job has the wrong type.  This means that there is no algorithmic 
    way of distinguishing between an unknown job ID and specifying an 
    ID for the wrong kind of Job.

  o An SRMInvalidRequestException implies SRM_INVALID_REQUEST as a 
    response.

  o Prior to the patch, if the SRM client supplies the ID for a Job of 
    the wrong type, then sometimes the code returned SRM_FAILED 
    (PutDone, StatusOfGetRequest, StatusOfPutRequest, 
    StatusOfLsRequest) in other places SRM_INVALID_REQUEST is returned 
    (StatusOfCopyRequest, ReleaseFiles, GetRequestSummary, 
    AbortRequest, ...).

    My reading of the SRM spec suggests that SRM_INVALID_REQUEST is the 
    correct response for all these circumstances.  This patch changes 
    the behaviour to return SRM_INVALID_REQUEST if the SRM client 
    supplies an ID that corresponds to the wrong kind of Job.

  o Both before and after the patch, getJob can never return null.

  o Despite this, the code contains tests for getJob returning null.  
    All such code is dead code-path.  This patch removes the test and 
    the dead code-path to avoid future confusion.

  o There are a few places where I was unsure if SRM relied on 
    side-effects of calling a method (and a few places where I'm sure 
    it does!).  Rather than attempting to fix these in this patch, I've 
    added a "FIXME" comment.  Future patches should tackle these.

Target: trunk
Patch: http://rb.dcache.org/r/3528/
Acked-by: Gerd Behrmann
Acked-by: Dmitry Litvintsev