non-LoginManager doors: faster registration with LoginBroker
The LoginBroker holds information about which doors are running within 
dCache.  Doors send messages to update the LoginBroker, both about the 
doors existence and the current load on the door.

During dCache startup, it is possible for doors to emit the first 
message to the LoginBroker before the LoginBroker has started accepting 
messages.  If this happens then LoginBroker will have to wait the 
duration of the update period before the door will appear in the 
LoginBroker.  The default value for this is five seconds.

The SRM uses the LoginBroker to establish which doors are running in 
dCache and, therefore, which protocols are available.  Since a door may 
be unknown to the LoginBroker on startup, the SRM can fail to issue 
TURLs during that period.  This has an adverse affect on the ability to 
test dCache.

This patch reduces the period where a door is unknown to the 
LoginBroker.  In effect, the door has two modes: normal and eager.  
When in eager-mode, the door will attempt to contact the LoginBroker 
every second.  In normal mode the door will attempt to contact 
LoginBroker using the usual policy (a configurable delay and checking 
by how much the load has changed).  An attempt to contact LoginBroker 
that fails switches LoginManager into eager-mode; communication that 
succeeds switches LoginManager into normal-mode.

This reduces the delay between either the LoginBroker or the door 
starting (whichever happens last) and the door being known to 
LoginBroker to, at most, one second (+ overheads like transport, 
de-/serialisation, etc).

Note that this patch is similar to a previous patch (SVN Trunk r.16004) 
that introduced the same behaviour to LoginManager-based doors.

Target: trunk
Patch: http://rb.dcache.org/r/3582/
Acked-by: Gerd Behrmann