ftp: Avoid error during unclean control channel shutdown
When the client disconnects in an unclean fashion, the following error
was logged by the FTP door:

19 Jul 2011 11:55:11 (GFTP-dcache-vm-<unknown>-103)
[door:GFTP-dcache-vm-<unknown>-103@gridftp-Domain:1311069271832
door:GFTP-dcache-vm-<unknown>-103@gridftp-Domain:1311069271832] Failed
to shut down input stream of the control channel: Transport endpoint is
not connected


The door tries to shut down the input stream of the control channel, but
it fails because the socket is already disconnected. There is no way to
avoid the exception as there is no way to query the state of the
connection; methods like isClosed() and isConnected() simply return the
state of the Socket object, not the connection.

The patch lowers the log level to avoid printing the error at the
default log level.

Target: trunk
Request: 1.9.13
Request: 1.9.12
Acked-by: Karsten Schwank <karsten.schwank@desy.de>
Patch: http://rb.dcache.org/r/3606/