ftp: Avoid NPE in info command
The info command of the FTP door throws an NPE if the user isn't logged
in yet.

The patch fixes the problem. It also reduces the throws declaration of
the ac_info method to ensure that a NPE would be logged.

Target: trunk
Request: 1.9.13
Request: 1.9.12
Acked-by: Karsten Schwank <karsten.schwank@desy.de>
Patch: http://rb.dcache.org/r/3604/