all: Let cells propagate declared exceptions back to client
Cells can deliver two types of messages: Strings are processed by ac_
methods and other messages by messageArrived methods. Checked exceptions
thrown by these methods are propagated back to the client cell as
failures. RuntimeExceptions (with the exception of a few) were treated
as bugs and logged as such in the cell.

The problem is that some methods throw RuntimeExceptions as part of
their contract. A typical example is throwing NoSuchElementException to
indicate that the requested element doesn't exist. In pools these would
show up as:

14 Jul 2011 19:00:04 (pool_0) [] Uncaught exception in thread pool-9-thread-5
java.util.NoSuchElementException: 232
        at org.dcache.pool.classic.PoolV4.ac_mover_ls_$_0_1(PoolV4.java:2004) ~[dcache.jar:na]
        at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method) ~[na:1.6.0_24]
        at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:39) ~[na:1.6.0_24]
        at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAcces
sorImpl.java:25) ~[na:1.6.0_24]
        at java.lang.reflect.Method.invoke(Method.java:597) ~[na:1.6.0_24]
        at dmg.util.CommandInterpreter.execute(CommandInterpreter.java:590) ~[cells.jar:na]
        at dmg.util.CommandInterpreter.command(CommandInterpreter.java:469) ~[cells.jar:na]
        at dmg.cells.nucleus.CellAdapter.executeLocalCommand(CellAdapter.java:1009) ~[cells.jar:na]
        at dmg.cells.nucleus.CellAdapter.messageArrived(CellAdapter.java:919) ~[cells.jar:na]
        at dmg.cells.nucleus.CellNucleus$DeliverMessageTask.innerRun(CellNucleus.java:1086) ~[cells.jar:na]
        at dmg.cells.nucleus.CellNucleus$AbstractNucleusTask.run(CellNucleus.java:984) ~[cells.jar:na]
        at java.util.concurrent.ThreadPoolExecutor$Worker.runTask(ThreadPoolExecutor.java:886) [na:1.6.0_24]
        at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:908) [na:1.6.0_24]
        at java.lang.Thread.run(Thread.java:662) [na:1.6.0_24]


This patch fixes the problem by looking at the exceptions declared
to be thrown by the message handler. All exceptions, even
RuntimeExceptions, declared to be thrown are now treated as
part of the public contract and silently propagated back to the
client. Other RuntimeExceptions (excluding IllegalArgumentException
and IllegalStateException for compatibility with existing code)
are treated as bugs and logged with a stack trace.

Target: trunk
Request: 1.9.13
Request: 1.9.12
Acked-by: Jan Schaefer <jan.schaefer@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3594/