srm, pool: Don't use delegation for GridFTP mover used for srmCopy
To implement srmCopy, the SRM uses the transfer manager to start a
RemoteGsiftpTransferProtocol mover. The mover submits a delegation
request back to the SRM, which the creates a TCP connection to the pool
and uses GSI to delegate the delegated credentials to the pool.

In general this is the correct way to pass credentials between services.
However in the case of dCache the two components really belong to the
same high level service and they communicate through what we must assume
is a safe channel (cells messaging; if not we have much bigger problems
in dCache).

Using GSI delegation comes at a high cost in terms of latency
(additional messages, additonal TCP connections) and resources (TCP
connections, CPU time - particularly a problem on the SRM server).

Therefore this patch replaces the above mechanism. Instead the delegated
credetials received from the user are embedded in the protocol info and
submitted to the pool.

The old mechanism is left in place for older pools. It can be removed
when we decide to break backwards compatibility.

Target: trunk
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3597/