ftp: Fix race condition in command execution
An FTP door uses a thread to read from the control channel and uses a different 
thread for command execution. This is necessary to detect that the control 
channel is closed while commands are executed.

Most of the time no command is executed and hence allocating a dedicated thread 
per FTP door for command execution is a waste of resources. Therefore FTP door 
instances within the same door use a shared Executor.

Commands from a particular session however need to be executed sequentially. 
For that reason the FTP door uses a CommandQueue to ensure command ordering and 
the dynamic submission of command execution tasks to the shared executor. Once 
the queue is empty the task shuts down. When another command is added a new 
execution task is submitted.

Unfortunately there is a race condition in the command execution task: It 
fetches commands by calling the synchronized get() method. If empty, null is 
returned. It then calls the synchronized done() to signal that the task shut 
down and that further commands being added must resubmit the execution task. 
Obviously there is a window between calling get() and done() in which new 
commands could be added to the queue without a new execution task being 
submitted.

The symptom would be that the FTP door appears to be stuck.

The fix is to combine get() and done() into a single synchronized method. The 
fix also resolves potential issues with RuntimeExceptions being thrown while 
executing an FTP command.

Target: trunk
Request: 1.9.13
Request: 1.9.12
Request: 1.9.11
Request: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3602/
Committed: trunk@16050
Merge-request: http://rt.dcache.org/Ticket/Display.html?id=6483
Signed-off-by: Gerd Behrmann