ftp: Allow GSI handshake to be interrupted
JGlobus has a race condition in which the GSI handshake will get stuck if an
incomplete SSL message is provided by the client. It is difficult to make a
workaround for that issue.

To make the situation worse, the FTP door doesn't interrupt the handshake if
the control channel is closed. The result is that even when the client
disconnects, the FTP door instance keeps running.

The patch fixes the problem by enabling the handshake to be interrupted.

Target: trunk  
Request: 1.9.13
Request: 1.9.12
Request: 1.9.11
Request: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3603/
Signed-off-by: Gerd Behrmann
Committed: trunk@16049