poolmanager: Don't save setup using legacy format
Back in the late 1.6 days the partition manager was introduced. It
allows various configuration settings to be specified per partition.

Partition manager continued to support reading and writing configuration
using the older format. The new format would only be used for
non-default partitions.

As long as we continue to write the setup using the legacy format we
cannot deprecate the legacy commands. As a first to get rid of the
legacy commands, this patch changes the pool manager to store its setup
using the "new" format.

I have to stress that the "new" format has been supported by dCache for
the last 5 years.

I propose to merge this change into 1.9.12 so that we can get rid of the
legacy commands in 1.9.14, thus paving the road for further pool manager
improvements.

Target: trunk
Request: 1.9.13
Request: 1.9.12
Acked-by: Karsten Schwank <karsten.schwank@desy.de>
Patch: http://rb.dcache.org/r/3650/