pool: Make proportional pool selection proportional to free space
The migration module has several pool selection strategies. The default
is proportional. It used to be proportional to the inverse of the space
cost. There have been several reports about uneven distribution of
files. That can happen when a pool is empty. In that case the space cost
is zero. The inverse is infinity, and thus the first such pool in the
list would attract all transfers for a while.

This patch modifies the proportional pool selection such that a pool is
selected with a probably proportional to the amount of free space. Thus
a pool with twice as much free space has twice the probability to be
selected.

Cached files are taken into account by applying an exponential decay. A
newly accessed cached file is accounted for as filling the actual size
of the file. As time passes without the file being accessed, it decays
("shrinks") exponentially during the pool selection. Thus as time goes
to infinity, all cacheable files would be considered free space. The
halflife is controlled by the breakeven parameter of the target pool.

Since the last access time isn't known for every cached file, a linear
extrapolation is used with the last access time of the LRU cached file
as the sole input.

The algorithm was proposed by Mattias Wadenstein a couple of years ago.
I see this implementation as a low risk trial for the algorithm. If it
works well we could adopt if for regular pool selection in PoolManager.

Although the change is relatively large for a bug fix, I still 
eventually submit if for merge into 1.9.12 and 1.9.13. The reason is 
that we have several reports complaining about the old algorithm and 
that I consider it low risk despite the large change to the algorithm.

Target: trunk
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6205
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6478
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/3634/