poolmanager: Rewrite PartitionManager
Pool manager uses PartitionManager to manage collections of properties.
This patch refactors (more or less a rewrite) PartitionManager to clean
up the code, improve efficiency and correctness. More importantly, the
patch is a foundation of a major restructuring of pool selection (more
about that below).

Some key differences to the old code:

- PoolManagerParameter got renamed into ClassicPartition, with Partition
  being a base class.

- Properties of Partition are immutable, meaning they are inheritenly thread
  safe. Thus the copying of PoolManagerParameter that happened all the time
  in the old code is gone.

- Partition follows a copy on write semantics, meaning the partitions have
  to be recreated for each update. Updates are however infrequent.

- While partitions are updated, the old partitions continue to be valid,
  meaning that pool manager can continue do what it does even when
  parameters are being updated.

- PartitionManager is now ignorant of the precise properties supported.
  All property specific logic is in Partition/ClassicPartition.

- The split into a base class and a subclass is a first step to different
  types of partitions. Thus partitions are no longer limited to define
  different values of a fixed set of properties. It is possible to
  define new types of partitions with new parameters. This isn't actually
  implemented yet, but the structure is ready for this.

Partitions may inherit from a common set of properties maintained by
PartitionManager. The 'pm ls' command indicates whether the current
value is a default value, an inherited value or a value defined
specifically for this partition.

'pm set' can define properties in this common set. For backwards
compatibility, 'pm set default' applies changes to this common set
rather than to the default partition. We may change this in the future,
but doing so will break compatibility with the poolmanager.conf
generated by versions 1.9.12-9 and newer.

The plan is to move the various types of pool selection logic into the
partition. The partitioning of the parameters between Partition and
ClassicPartition is with this next step in mind: The parameters in
ClassicPartition are specific to our current pool selection logic. Other
pool selection strategies may need other parameters. The parameters in
Partition on the other hand are used by pool manager itself and are not
strictly related to pool selection.

Currently pool selection is unchanged. For this reason properties of
Partition and ClassicPartition are public. This will change, but I decided
to keep like this in this patch to avoid inflating the patch further.

Requires synchronous update of head nodes.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3654/