poolmanager: Make partition type selectable and pluggable
A PartitionFactory interface is introduced. Java's ServiceLoader
mechanism is used to locate implementations.

PartitionManager is extended to allow the partition type to be specified
for new partitions. The type of the default partition is hard coded (may
be changed in the future).

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3667/