Webadmin: Letting the spacetokens page use the pagecache instead of info-xml
Rather big patch. It includes several changes that were necessary to change the datasource from info to cell-communication. Especcially the change of the datatypes made it neccessary to change a lot of classes. But all together it is a lot of mapping.

Patch: http://rb.dcache.org/r/3673/

Acked-by: Tigran

Target: trunk