Topo: Adding possibillity to list all hostnames in the dcache-instance to the topo-cell.
This is a requested feature of the DESY-admins in order not to need to have a list of all hosts outside dCache when you want to upgrade an instance and make sure not to forget a machine. This way you can retrieve this list from dCache itself. In a later patch I add this list to the Webadmin for admins.

Moved the NetworkUtils class into commons to be able to use it <-- separated out into another patch http://rb.dcache.org/r/3687/

Made an own Bean out of the added functionality for topo.

Patch: http://rb.dcache.org/r/3671/

Acked-By: Gerd and Tigran

Target: trunk
Request: 1.9.13
Request: 1.9.12