shell: remove leading and trailing whitespace around a property value
In dCache configuration, we say we ignore leading and trailing 
white-space of a property value; for example, the two lines "foo= bar " 
and "foo=bar" are equivalent.  When compiling the Shell Oracle, leading 
and trailing white-space is not removed.  The value is returned so that 
white-space is preserved; therefore, there is a distinction between the 
two lines above.

This issue triggered a problem when a sysadmin attempted to configure a 
domain that starts as user "root".  If they use a line like "user=root " 
then the startup scripts will attempt to change user to the non-existing 
user "root ".

Note: we should update the documentation to make it clear that we (now) 
strip trailing white-space.

Target: trunk
Request: 1.9.12
Request: 1.9.13
Patch: http://rb.dcache.org/r/3701/
Acked-by: Gerd Behrmann